refactor: isolate sidebar selection reducer
